test: move the https req.socket TLSSocket coverage to its own file
node-tls-server.test.ts has an environment-sensitive SNI test (binds and
connects via the localhost hostname) that fails in some containers
regardless of this change; a dedicated file keeps this fix's coverage
isolated from it.
