+++ test/js/node/tls/node-tls-server.test.ts
@@ -838,71 +838,6 @@ it("keeps req.socket.authorized false for an unverified client after the server 
   }
 });
 
-it("https req.socket is a TLSSocket that exposes the client certificate", async () => {
-  // https://github.com/oven-sh/bun/issues/37251: the socket node:https hands
-  // the request handler must be a tls.TLSSocket so mTLS servers can read the
-  // client identity via req.socket.getPeerCertificate().
-  const fixtures = join(import.meta.dir, "fixtures");
-  const serverOptions = {
-    key: readFileSync(join(fixtures, "agent10-key.pem"), "utf8"),
-    cert: readFileSync(join(fixtures, "agent10-cert.pem"), "utf8"),
-    ca: readFileSync(join(fixtures, "ca5-cert.pem"), "utf8"),
-    requestCert: true,
-    rejectUnauthorized: false,
-  };
-  const { promise, resolve, reject } = Promise.withResolvers<Record<string, unknown>>();
-  const server = https.createServer(serverOptions, (req, res) => {
-    try {
-      const s = req.socket as TLSSocket;
-      resolve({
-        ctor: s.constructor.name,
-        isTLSSocket: s instanceof TLSSocket,
-        isNetSocket: s instanceof net.Socket,
-        authorized: s.authorized,
-        peerCN: s.getPeerCertificate()?.subject?.CN,
-        detailedCN: s.getPeerCertificate(true)?.subject?.CN,
-        x509Subject: s.getPeerX509Certificate()?.subject,
-        cipherName: s.getCipher()?.name,
-        protocol: s.getProtocol(),
-      });
-    } catch (err) {
-      reject(err);
-    }
-    res.end("ok");
-  });
-  server.on("error", reject);
-  server.listen(0, "127.0.0.1");
-  await once(server, "listening");
-  const { port } = server.address() as AddressInfo;
-  const clientRequest = https.request({
-    port,
-    host: "127.0.0.1",
-    rejectUnauthorized: false,
-    key: readFileSync(join(fixtures, "ec10-key.pem"), "utf8"),
-    cert: readFileSync(join(fixtures, "ec10-cert.pem"), "utf8"),
-    agent: false,
-  });
-  clientRequest.on("error", reject);
-  clientRequest.end();
-  try {
-    const got = await promise;
-    expect(got).toEqual({
-      ctor: "TLSSocket",
-      isTLSSocket: true,
-      isNetSocket: true,
-      authorized: true,
-      peerCN: "agent10.example.com",
-      detailedCN: "agent10.example.com",
-      x509Subject: expect.stringContaining("CN=agent10.example.com"),
-      cipherName: expect.any(String),
-      protocol: expect.stringMatching(/^TLSv/),
-    });
-  } finally {
-    clientRequest.destroy();
-    server.close();
-  }
-});
-
 it("createServer({pfx, requestCert}) verifies client certificates against the pfx-embedded CA", async () => {
   // agent1.pfx bundles agent1's key/cert plus ca1; a server built from it must
   // be able to verify a client certificate signed by that embedded CA.
